+++ src/files.js
@@ -1,11 +1,10 @@
-'use strict';
-
 const Task = require('./task');
 
 class Files extends Task {
     constructor(zume, options) {
         options = Object.assign({
-            pattern: '**'
+            pattern: '**',
+            incremental: true
         }, options);
 
         options.watchPattern = options.watchPattern || options.pattern;