+++ src/plugins/yaml.js
@@ -1,11 +1,7 @@
-'use strict';
-
 const through = require('through2');
 const yaml = require('js-yaml');
 
-module.exports = function (options) {
-    options = options || {};
-
+module.exports = function (options = {}) {
     function run (file, done) {
         try {
             const data = yaml.load(file.contents.toString());