+++ src/img.js
@@ -1,19 +1,14 @@
 const Task = require('./task');
+const merge = require('merge-options');
+const defaults = {
+    base: 'img',
+    pattern: '**/*.{jpg,jpeg,png,svg,gif}',
+    incremental: true
+};
 
 class Img extends Task {
     constructor(zume, options) {
-        options = Object.assign(
-            {
-                base: 'img',
-                pattern: '**/*.{jpg,jpeg,png,svg,gif}',
-                incremental: true
-            },
-            options
-        );
-
-        options.watchPattern = options.watchPattern || options.pattern;
-
-        super(zume, options);
+        super(zume, merge(defaults, options));
         this.reload = ['*.jpg', '*.jpeg', '*.png', '*.svg', '*.gif'];
     }
 }