+++ src/js.js
@@ -1,17 +1,14 @@
 const Task = require('./task');
+const merge = require('merge-options');
+const defaults = {
+    base: 'js',
+    watchPattern: '**/*.js',
+    pattern: '*.js'
+};
 
 class Js extends Task {
     constructor(zume, options) {
-        options = Object.assign(
-            {
-                base: 'js',
-                watchPattern: '**/*.js',
-                pattern: '*.js'
-            },
-            options
-        );
-
-        super(zume, options);
+        super(zume, merge(defaults, options));
         this.reload = '*.js';
     }