+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [UNRELEASED]
+## [1.0.0-beta.3] - 2018-03-19
+
+### Fixed
+
+- Gulp error events are not handled.
+
+## [1.0.0-beta.2] - 2018-03-19
 
 ### Added
 
@@ -269,7 +275,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 First version
 
-[UNRELEASED]: https://github.com/oscarotero/zume/compare/v1.0.0-beta.1...HEAD
+[1.0.0-beta.3]: https://github.com/oscarotero/zume/compare/v1.0.0-beta.2...v1.0.0-beta.3
+[1.0.0-beta.2]: https://github.com/oscarotero/zume/compare/v1.0.0-beta.1...v1.0.0-beta.2
 [1.0.0-beta.1]: https://github.com/oscarotero/zume/compare/v0.15.1...v1.0.0-beta.1
 [0.15.1]: https://github.com/oscarotero/zume/compare/v0.15.0...v0.15.1
 [0.15.0]: https://github.com/oscarotero/zume/compare/v0.14.6...v0.15.0